refactor: 숫자 생성 Lotto -> LottoGenerator로 이동
