+++ src/main/java/lotto/model/LottoNumber.java
@@ -3,10 +3,9 @@ package lotto.model;
 import java.util.Objects;
 
 public class LottoNumber {
-    private static final String ERROR_OUT_OF_RANGE_BONUS_NUMBER = "볼 번호가 1~45 범위 내에 해당하지 않습니다.";
     static final int MIN_LOTTO_NUMBER = 1;
     static final int MAX_LOTTO_NUMBER = 45;
-
+    private static final String ERROR_OUT_OF_RANGE_BONUS_NUMBER = "볼 번호가 1~45 범위 내에 해당하지 않습니다.";
     private final Integer number;
 
     public LottoNumber(Integer number) {
